requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
psycopg2-binary
streamlit
pytest
python-dotenv